add get subscription and update subscription to backend
