+++ backend/models/WCACompAnnouncementSubscription.go
@@ -1,0 +1,13 @@
+package models
+
+type WCACompAnnouncementSubscriptions struct {
+	Id          int    `json:"-"`
+	CountryId   string `json:"countryId"`
+	CountryName string `json:"countryName"`
+	Subscribed  bool   `json:"subscribed"`
+}
+
+type UpdateWCAAnnouncementSubscriptionsRequestBody struct {
+	CountryName string `json:"countryName"`
+	Subscribed  bool   `json:"subscribed"`
+}
